// src/app/store/shop/shop.action.types.ts
export enum ShopActionTypes {
  ADD_PRODUCT_TO_CART = 'ADD_PRODUCT_TO_CART',
  REMOVE_PRODUCT_FROM_CART = 'REMOVE_PRODUCT_FROM_CART',
  INCREMENT_CART_ITEM_QUANTITY = 'INCREMENT_CART_ITEM_QUANTITY',
  DECREMENT_CART_ITEM_QUANTITY = 'DECREMENT_CART_ITEM_QUANTITY',
  CHECKOUT_CART_ITEM = 'CHECKOUT_CART_ITEM',
  LOAD_PRODUCT = 'LOAD_PRODUCT',
  LOAD_PRODUCT_SUCCESS = 'LOAD_PRODUCT_SUCCESS',
  LOAD_PRODUCT_FAILURE = 'LOAD_PRODUCT_FAILURE'
}